# Optimize columns deserialization